+++ src/client/index.js
@@ -7,6 +7,7 @@ import Router from 'react-router/lib/Router';
 import browserHistory from 'react-router/lib/browserHistory';
 import match from 'react-router/lib/match';
 import routes from '../shared/routes';
+import { IS_HOT_DEVELOPMENT } from '../shared/config';
 
 // Get the DOM Element that will host our React application.
 const container = document.querySelector('#app');
@@ -45,7 +46,7 @@ function renderApp() {
 }
 
 // The following is needed so that we can hot reload our App.
-if (process.env.NODE_ENV === 'development' && module.hot) {
+if (IS_HOT_DEVELOPMENT) {
   // Accept changes to this file for hot reloading.
   module.hot.accept('./index.js');
   // Any changes to our routes will cause a hotload re-render.