+++ src/server/clientAssets/index.js
@@ -1,26 +1,11 @@
 // This file resolves the assets available from our client bundle.
 
 import fs from 'fs';
-import path from 'path';
-import appRoot from 'app-root-path';
+import { CLIENT_BUNDLE_ASSETSJSON_FILEPATH } from '../config';
 
-const appRootPath = appRoot.toString();
-
-const assetsPath = path.resolve(
-  appRootPath,
-  process.env.CLIENT_BUNDLE_OUTPUT_PATH,
-  process.env.CLIENT_BUNDLE_ASSETS_FILENAME
+const ClientBundleAssets = JSON.parse(
+  fs.readFileSync(CLIENT_BUNDLE_ASSETSJSON_FILEPATH, 'utf8')
 );
-
-if (!fs.existsSync(assetsPath)) {
-  throw new Error(
-    `We could not find the "${process.env.CLIENT_BUNDLE_ASSETS_FILENAME}" file ` +
-    'containing a list of the assets of the client bundle.  Please ensure that ' +
-    'the client bundle has been built before the server bundle.'
-  );
-}
-
-const ClientBundleAssets = JSON.parse(fs.readFileSync(assetsPath, 'utf8'));
 
 // Convert the assets json it into an object that contains all the paths to our
 // javascript and css files.  Doing this is required as for production