+++ src/server/htmlPage/render.js
@@ -4,6 +4,7 @@ import { renderToString } from 'react-dom/server';
 import serialize from 'serialize-javascript';
 import Helmet from 'react-helmet';
 import clientAssets from '../clientAssets';
+import favicon from './favicon.ico';
 
 // :: [String] -> [String]
 function styleTags(styles) {
@@ -59,7 +60,7 @@ function render(rootReactElement : ?$React$Element, initialState : ?Object) {
         <meta charSet='utf-8' />
         <meta httpEquiv='X-UA-Compatible' content='IE=edge' />
         <meta httpEquiv='Content-Language' content='en' />
-        <link rel='shortcut icon' type='image/x-icon' href='/public/favicon.ico' />
+        <link rel='shortcut icon' type='image/x-icon' href='${favicon}' />
 
         ${helmet ? helmet.title.toString() : ''}
         ${helmet ? helmet.meta.toString() : ''}