+++ src/shared/routes/index.js
@@ -4,8 +4,9 @@ import React from 'react';
 import Route from 'react-router/lib/Route';
 import IndexRoute from 'react-router/lib/IndexRoute';
 import App from '../components/App';
+import { IS_HOT_DEVELOPMENT } from '../config';
 
-if (process.env.NODE_ENV === 'development' && module.hot) {
+if (IS_HOT_DEVELOPMENT) {
   // HMR falls over when a Route uses a dynamic component resolution
   // property (i.e. getComponent or getComponents).  As a workaround for any
   // of your components that are resolved dynamically please require them below.